backend/requirements_frozen.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.12.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.2
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
duckduckgo_search==6.3.7
faiss-gpu-cu12==1.13.0
Farama-Notifications==0.0.4
fastapi==0.115.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
grpcio==1.76.0
gymnasium==0.29.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
ollama==0.3.3
packaging==25.0
pandas==2.3.3
pillow==12.0.0
primp==0.15.0
protobuf==6.33.1
psutil==7.1.3
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.26.6
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-triton==3.5.1+gitbfeb0668
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==3.1.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stable_baselines3==2.3.2
starlette==0.38.6
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.10.0.dev20251122+cu128
torchaudio==2.10.0.dev20251122+cu128
torchvision==0.25.0.dev20251122+cu128
tqdm==4.66.5
transformers==4.57.1
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
